Refactored auth0 getting
