+++ src/API/src/review/review.service.ts
@@ -1,28 +1,9 @@
 import { MailerService } from '@nestjs-modules/mailer';
-import { HttpService } from '@nestjs/axios';
 import { HttpException, Injectable, Logger } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
-import { lastValueFrom, map } from 'rxjs';
+import { AuthService } from 'src/auth/auth.service';
 import { Dataset } from 'src/db/shared/entities/dataset.entity';
 import { Repository } from 'typeorm';
-
-export const getAuth0Token = async (http: HttpService) => {
-  const token = await lastValueFrom(
-    http.post(`https://dev-326tk4zu.us.auth0.com/oauth/token`, {
-      grant_type: 'client_credentials',
-      client_id: 'sQPoZzmH4QaAHVEJrDaK3pPeHG0SmCtr',
-      client_secret: 'u6gsaL3kmxHPSmAYVXoJxaD5gbz0x3WhyRZQAzY7calz0y40rwmP2wOBbzigVTOt',
-      audience: 'https://dev-326tk4zu.us.auth0.com/api/v2/'
-    }, {headers: {'content-type': 'application/x-www-form-urlencoded',
-       "Accept-Encoding": "gzip,deflate,compress"}}).pipe(
-      map((res: any) => {
-        return res.data.access_token;
-      }),
-    ),
-  );
-  return token;
-};
-
 
 @Injectable()
 export class ReviewService {
@@ -31,12 +12,11 @@ export class ReviewService {
         private datasetRepository: Repository<Dataset>,
         private logger: Logger,
         private readonly mailerService: MailerService,
-        private readonly httpService: HttpService,
+        private readonly authService: AuthService,
     ){}
 
     async reviewDataset(datasetId: string, reviewerId?:string, reviewFeedback?: string){
       try{
-        const token = await getAuth0Token(this.httpService);
        if(datasetId){
           const dataset = await this.datasetRepository.findOne({
              where: {id: datasetId}
@@ -49,15 +29,7 @@ export class ReviewService {
 
            const userId = dataset.UpdatedBy;
 
-           const emailAddress = await lastValueFrom(
-            this.httpService.get(`https://dev-326tk4zu.us.auth0.com/api/v2/users/${userId}`,
-            {headers: {authorization: `Bearer ${token}`, "Accept-Encoding": "gzip,deflate,compress"}})
-            .pipe(
-              map((res: any) => {
-                return res.data.email;
-              }),
-            ),
-          );
+           const emailAddress = await this.authService.getEmailFromUserId(userId);
 
 
            const review_res = `<div>
@@ -113,26 +85,9 @@ export class ReviewService {
       );
 
       const uploader= dataset.UpdatedBy;
-      const token = await getAuth0Token(this.httpService);
 
-      const uploaderEmail = await lastValueFrom(
-       this.httpService.get(`https://dev-326tk4zu.us.auth0.com/api/v2/users/${uploader}`,
-       {headers: {authorization: `Bearer ${token}`, "Accept-Encoding": "gzip,deflate,compress"}})
-       .pipe(
-         map((res: any) => {
-           return res.data.email;
-         }),
-       ),
-     );
-      const approverEmail = await lastValueFrom(
-       this.httpService.get(`https://dev-326tk4zu.us.auth0.com/api/v2/users/${userId}`,
-       {headers: {authorization: `Bearer ${token}`, "Accept-Encoding": "gzip,deflate,compress"}})
-       .pipe(
-         map((res: any) => {
-           return res.data.email;
-         }),
-       ),
-     );
+      const uploaderEmail = await this.authService.getEmailFromUserId(uploader);
+      const approverEmail = await this.authService.getEmailFromUserId(userId);
 
      let approvalText = `<div>
            <h2>Data Approval</h2>
